#41 Store categories in the user state
